move VarArgHelper from domain to logic
and make constructor of ArgumentMapperBuilder internal
