+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/creating/Pairs.kt
@@ -1,6 +1,6 @@
 package ch.tutteli.atrium.api.infix.en_GB.creating
 
-import ch.tutteli.atrium.domain.builders.utils.VarArgHelper
+import ch.tutteli.atrium.logic.utils.VarArgHelper
 
 /**
  * Parameter object to express `Pair<K, V>, vararg Pair<K, V>`.
